Move login UI to index route
